+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Changes
 
 - [BREAKING] Incremented MSRV to 1.89.
+- Added `AccountComponent::from_package()` method to create components from `miden-mast-package::Package`.
 
 ## 0.11.0 (2025-08-26)